% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatVal.R
\name{statval}
\alias{statval}
\title{Validate a Partition}
\usage{
statval(df, targetVar, split, pval)
}
\arguments{
\item{df}{The data frame to be partitioned}

\item{targetVar}{The target variable of data frame entered as df$target}

\item{split}{Specified percent of data for training set}

\item{pval}{Specified p-value for statistical tests}
}
\value{
List of training and testing partitions along with p-values
}
\description{
Function creates a train and test partition with specified split percentage
and then validates the partition across all variables.
}
\details{
Train and Test data frames are output via a list--along with p-values--and can
be unpacked using their respective unpack functions.
}
