% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatVal.R
\name{unpack.test}
\alias{unpack.test}
\title{Unpack the Testing Data Set}
\usage{
unpack.test(statVal.saved)
}
\arguments{
\item{statVal.saved}{User defined name given to the return of statval}
}
\value{
Data frame of the validated testing set
}
\description{
Unpacks the testing set as data frame resulting from the statval function
}
