% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatVal.R
\name{unpack.train}
\alias{unpack.train}
\title{Unpack the Training Data Set}
\usage{
unpack.train(statVal.saved)
}
\arguments{
\item{statVal.saved}{User defined name given to the return of statval}
}
\value{
Data frame of the validated training set
}
\description{
Unpacks the training set as data frame resulting from the statval function
}
